remove grid_iou at inference. restore min_annot_1 in sample folder name. fix circular imports with read_csv() pass Travis CI. update requirements.txt
