+++ conf/config_ci_segmentation_local.yaml
@@ -26,14 +26,14 @@ sample:
   prep_csv_file: ./data/images_to_samples_ci_csv.csv
   overlap: 33
   val_percent: 30 # Percentage of validation samples created from train set (0 - 100)
-  sampling: {'method':['min_annotated_percent'], 'map': 1, '0':0, '1':0, '2':0} # One of class_proportion, annotated_percent
-      # 'method' : One or both in any order of min_annotated_percent, class_proportion; must be presented in list format
-      # 'map' : Min % of non background pixels in samples. Default: 0
-      # following keys:
-      # keys must include all classes present in the images
-      # keys must be numerical values in 'string' format
-    # class value must be in 'integer' format
-    # class value represents class minimum threshold targeted in samples
+  sampling_method: {'min_annotated_percent':0} # One of class_proportion, annotated_percent
+    # 'method' : One or both in any order of min_annotated_percent, class_proportion; must be presented in list format
+    # 'map' : Min % of non background pixels in samples. Default: 0
+    # following keys:
+    # keys must include all classes present in the images
+    # keys must be numerical values in 'string' format
+  # class value must be in 'integer' format
+  # class value represents class minimum threshold targeted in samples
   mask_reference: False
 
 # Data analysis parameters; used in data_analysis.py ------------------
@@ -59,6 +59,7 @@ training:
   num_tst_samples:
   batch_size: 4
   num_epochs: 2
+  target_size: 256
   loss_fn: Lovasz # One of CrossEntropy, Lovasz, Focal, OhemCrossEntropy (*Lovasz for segmentation tasks only)
   optimizer: adam # One of adam, sgd or adabound
   learning_rate: 0.0001
@@ -82,7 +83,7 @@ training:
 
 inference:
   img_dir_or_csv_file: ./data/inference_sem_seg_ci_csv.csv
-  state_dict_path: ./data/samples256_overlap33_min-annot1_3bands/model/config_ci_segmentation_local/checkpoint.pth.tar
+  state_dict_path: ./data/samples256_overlap33_min-annot0_3bands/model/config_ci_segmentation_local/checkpoint.pth.tar
   chunk_size: 512 # (int) Size (height and width) of each prediction patch. Default: 512
   overlap: 10 # (int) Percentage of overlap between 2 chunks. Default: 10
   heatmaps: True
